+++ source/simulation2/components/CCmpPosition.cpp
@@ -395,7 +395,7 @@ public:
 			float rotY = m_RotY.ToFloat();
 			float delta = rotY - m_InterpolatedRotY;
 			// Wrap delta to -M_PI..M_PI
-			delta = fmod(delta + (float)M_PI, 2*(float)M_PI); // range -2PI..2PI
+			delta = fmodf(delta + (float)M_PI, 2*(float)M_PI); // range -2PI..2PI
 			if (delta < 0) delta += 2*(float)M_PI; // range 0..2PI
 			delta -= (float)M_PI; // range -M_PI..M_PI
 			// Clamp to max rate
